사진이 n개에서 0개로 바꿀 경우, 썸네일 제거

+++ app/src/main/java/hbs/com/linememo/ui/memo_make/MemoMakeActivity.kt
@@ -87,7 +87,7 @@ class MemoMakeActivity : BaseActivity() {
                     compositeDisposable.add(memoMakeViewModel.insertMemo(this)
                         .flatMap { memoId ->
                             if (memoMakeGalleryAdapter.currentList.size > 0) {
-                                memoMakeViewModel.insertMemoGalleries(memoId, memoMakeGalleryAdapter.currentList)
+                                memoMakeViewModel.insertMemoGalleries(memoId.toInt(), memoMakeGalleryAdapter.currentList)
                             }else{
                                 Observable.just(0)
                             }
